Fux Bug

+++ app/controllers/items_controller.rb
@@ -26,7 +26,7 @@ class ItemsController < ApplicationController
     # @item.name = @name
     # @item.category = @category
     # @item.enable = false
-    # @item.picture.attach(params[:picture])        # PROBLEM: Picture doesn't show
+    # @item.picture.attach(params[:picture])      
     # @item.save
 
     # @market = Market.new
@@ -53,15 +53,25 @@ class ItemsController < ApplicationController
   def new
     @item = Item.new
     @editing = false
+    if (@myrole == 'buyer')
+      redirect_to '/permission'
+    end
   end
 
   # GET /items/1/edit
   def edit
     @editing = true
+    @belong_user_id = Market.where(item_id: params[:id]).first.user_id
+    if (@myrole != 'admin' && @belong_user_id != session[:user_id])
+      redirect_to '/permission'
+    end
   end
 
   # POST /items or /items.json
   def create
+    if (@myrole == 'buyer')
+      redirect_to '/permission'
+    end
     @editing = false
     @item = Item.new(item_params)
 
@@ -90,6 +100,9 @@ class ItemsController < ApplicationController
 
   # PATCH/PUT /items/1 or /items/1.json
   def update
+    if (@myrole == 'buyer')
+      redirect_to '/permission'
+    end
     respond_to do |format|
       if @item.update(item_params)
         if (item_params[:picture] != nil)
@@ -106,6 +119,9 @@ class ItemsController < ApplicationController
 
   # DELETE /items/1 or /items/1.json
   def destroy
+    if (@myrole == 'buyer')
+      redirect_to '/permission'
+    end
     Inventory.where(item_id: @item.id).destroy_all
     Market.where(item_id: @item.id).destroy_all
     @item.destroy
